predibench-backend/requirements.txt
gradio
datasets
huggingface_hub
apscheduler
python-dotenv
git+https://github.com/aymeric-roucher/PrediBench#subdirectory=predibench-core